requirements.txt
# pyjson5
packaging
jinja2
arpy
mistune
pillow